Run webpack only on server mode

+++ config/initializers/webpack.rb
@@ -1,11 +1,15 @@
 if Rails.env.development?
   Rails.application.config.before_initialize do
-    puts '=> Starting Webpack in watch mode'
-    `cd sequencer-react && npm run dev-rails > ../log/webpack.log & echo $! > tmp/pids/webpack.pid`
+    if defined?(Rails::Server)
+      puts '=> Starting Webpack in watch mode'
+      `cd sequencer-react && npm run dev-rails > ../log/webpack.log & echo $! > tmp/pids/webpack.pid`
+    end
   end
 
   at_exit do
-    puts "Exiting Webpack"
-    `/bin/kill -s SIGINT $(cat tmp/pids/webpack.pid) && rm tmp/pids/webpack.pid`
+    if defined?(Rails::Server)
+      puts "Exiting Webpack"
+      `/bin/kill -s SIGINT $(cat tmp/pids/webpack.pid) && rm tmp/pids/webpack.pid`
+    end
   end
 end
